[#78] :: Rewrite old Board component logic
